requirements.txt
flask
flask-cors
python-dotenv
requests
edge-tts
